requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.1
pillow==9.5.0
numpy==1.25.1
cairosvg==2.7.0
python-dotenv==1.0.0
